Modelando área geral

+++ Prova/vacinacao/app/Http/Controllers/PessoaController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Pessoa;
+use Illuminate\Support\Facades\Auth;
 
 class PessoaController extends Controller
 {
@@ -25,13 +26,18 @@ class PessoaController extends Controller
     public function store(Request $request)
     {
         try {
-            Pessoa::create([
-                'nome' => $request->nome,
-                'bairro' => $request->bairro,
-                'cidade' => $request->cidade,
-                'data_nascimento' => $request->data_nascimento,
-            ]);
-            $request->session()->flash('success', 'Pessoa cadastrado com sucesso.');
+            if (Auth::user()) {
+                Pessoa::create([
+                    'nome' => $request->nome,
+                    'bairro' => $request->bairro,
+                    'cidade' => $request->cidade,
+                    'data_nascimento' => $request->data_nascimento,
+                ]);
+                $request->session()->flash('success', 'Pessoa cadastrado com sucesso.');
+            } else {
+                $request->session()->flash('warning', 'Você não possui permissão para executar esta ação.');
+                return redirect()->back();
+            }
         } catch (\Throwable $th) {
             report($th);
             $request->session()->flash('error', 'Erro ao cadastrar pessoa, tente novamente.');
@@ -52,13 +58,18 @@ class PessoaController extends Controller
     public function update(Request $request, $id)
     {
         try {
-            Pessoa::findOrFail($id)->update([
-                'nome' => $request->nome,
-                'bairro' => $request->bairro,
-                'cidade' => $request->cidade,
-                'data_nascimento' => $request->data_nascimento,
-            ]);
-            $request->session()->flash('success', 'Pessoa atualizado com sucesso.');
+            if (Auth::user()) {
+                Pessoa::findOrFail($id)->update([
+                    'nome' => $request->nome,
+                    'bairro' => $request->bairro,
+                    'cidade' => $request->cidade,
+                    'data_nascimento' => $request->data_nascimento,
+                ]);
+                $request->session()->flash('success', 'Pessoa atualizado com sucesso.');
+            } else {
+                $request->session()->flash('warning', 'Você não possui permissão para executar esta ação.');
+                return redirect()->back();
+            }
         } catch (\Throwable $th) {
             report($th);
             $request->session()->flash('error', 'Erro ao atualizar pessoa, tente novamente.');
@@ -69,8 +80,13 @@ class PessoaController extends Controller
     public function destroy(Request $request, $id)
     {
         try {
-            Pessoa::findOrFail($id)->delete();
-            $request->session()->flash('sucess', 'Pessoa deletada com sucesso.');
+            if (Auth::user()) {
+                Pessoa::findOrFail($id)->delete();
+                $request->session()->flash('sucess', 'Pessoa deletada com sucesso.');
+            } else {
+                $request->session()->flash('warning', 'Você não possui permissão para executar esta ação.');
+                return redirect()->back();
+            }
         } catch (\Throwable $th) {
             report($th);
             $request->session()->flash('error', 'Erro ao deletar pessoa, tente novamente');
